Changed to better looking d3-slider library.
The functionality is all totally the same, but now we don't have to worry
about the slider looks or the slider.js script. I've put an example of how
we could build the string for selecting the correct tree, but the generate
tree script likely needs to be changed so it allows the tree to be
updated. It should be easy to connect the images to the right dataset once
the datasets exist.